+++ gulpfile.js
@@ -61,7 +61,8 @@ var compilerMinimalist = webpack({
   plugins: [
     bannerPlugin,
     new webpack.IgnorePlugin(new RegExp('^brace$')),
-    new webpack.IgnorePlugin(new RegExp('^ajv'))
+    new webpack.IgnorePlugin(new RegExp('^ajv')),
+    new webpack.IgnorePlugin(new RegExp('^vanilla-picker$'))
   ],
   cache: true
 });